rtti 다양한 인자를 갖는 생성자 지원

+++ SMGE/Components/CDrawComponent.h
@@ -37,7 +37,7 @@ namespace SMGE
 		virtual SGReflection& operator=(CVector<TupleVarName_VarType_Value>& in) override;
 
 		SGRefl_Transform sg_transform_;
-		mutable int32_t persistentComponentNumber_ = 0;
+		mutable size_t persistentComponentNumber_ = 0;
 
 		TReflectionClass& outerDrawCompo_;
 	};
